requirements.txt
aioschedule==0.5.2
google-api-python-client==2.183.0
httplib2==0.31.0
oauth2client==4.1.3
python-telegram-bot==22.4
python-dotenv==1.1.1
schedule==1.2.2
sqlalchemy==2.0.43
